Updated architecture to support any number of output classes
