+++ lib/fastlane_core/ui/implementations/terminal.rb
