+++ lib/fastlane_core/ui/ui.rb
@@ -19,23 +19,10 @@ module FastlaneCore
 end
 
 require 'fastlane_core/ui/interface'
-require 'fastlane_core/ui/terminal'
 
-# Disabling colors
-
-if ENV["FASTLANE_DISABLE_COLORS"]
-  require 'colored'
-
-  class String
-    Colored::COLORS.keys.each do |color|
-      define_method(color) do
-        self # do nothing with the string, but return it
-      end
-    end
-    Colored::EXTRAS.keys.each do |extra|
-      define_method(extra) do
-        self # do nothing with the string, but return it
-      end
-    end
-  end
+# Import all available implementations
+Dir[File.expand_path('implementations/*.rb', File.dirname(__FILE__))].each do |file|
+  require file
 end
+
+require 'fastlane_core/ui/disable_colors' if ENV["FASTLANE_DISABLE_COLORS"]
